# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/bookmark/{id}": {
            "get": {
                "summary": "Display the specified resource.",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessBioController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Business uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business bio"
                    }
                }
            }
        },
        "/api/v1/bookmark": {
            "put": {
                "summary": "Updates a business bio",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessBioController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Business ID",
                                        "type": "string"
                                    },
                                    "bio": {
                                        "description": "Business bio",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "summary": "Toggle bookmark specified by ID for logged in user",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::toggleBookmark",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookmark successfully created/deleted!"
                    }
                }
            }
        },
...
        "/api/v1/business-hours/{business}": {
            "put": {
                "summary": "Update business hours",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessHoursController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "open_period_mins": {
                                        "description": "Start time",
                                        "type": "string"
                                    },
                                    "close_period_mins": {
                                        "description": "End time",
                                        "type": "string"
                                    },
                                    "business_id": {
                                        "description": "ID of business",
                                        "type": "string"
                                    },
                                    "wd_0": {
                                        "description": "Sunday",
                                        "type": "string"
                                    },
                                    "wd_1": {
                                        "description": "Monday",
                                        "type": "string"
                                    },
                                    "wd_2": {
                                        "description": "Tuesday",
                                        "type": "string"
                                    },
                                    "wd_3": {
                                        "description": "Wednesday",
                                        "type": "string"
                                    },
                                    "wd_4": {
                                        "description": "Thursday",
                                        "type": "string"
                                    },
                                    "wd_5": {
                                        "description": "Friday",
                                        "type": "string"
                                    },
                                    "wd_6": {
                                        "description": "Saturday",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "BusinessHoursResource"
                    }
                }
            }
        },
...
    }
}